% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_dispersal.R
\name{fun_dispersal}
\alias{fun_dispersal}
\title{Internal function: dispersal}
\usage{
fun_dispersal(x, m_p_dispersal, m_dispersal, boundary_condition)
}
\arguments{
\item{x}{Matrix of population sizes}

\item{m_p_dispersal}{Matrix of dispersal probability}

\item{m_dispersal}{Dispersal matrix}

\item{boundary_condition}{Define boundary condition for dispersal. \code{retain} has not loss, \code{loss} induces net loss out of the network.}
}
\description{
Internal function: dispersal
}
\author{
Akira Terui, \email{hanabi0111@gmail.com}
}
