% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_dyn_set.R
\name{fun_dyn_set}
\alias{fun_dyn_set}
\title{Internal function: function for community dynamics}
\usage{
fun_dyn_set(type)
}
\arguments{
\item{type}{Function type.}
}
\description{
Internal function: function for community dynamics
}
