% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_r_tmp.R
\name{fun_r_tmp}
\alias{fun_r_tmp}
\title{Internal function: function for temperature-dependent population growth}
\usage{
fun_r_tmp(
  a,
  e_a,
  e_d,
  k_b = 8.6 * 1e-05,
  tmp,
  tmp_r = 298.15,
  tmp_h,
  scale_factor = 0
)
}
\arguments{
\item{a}{normalization constant.}

\item{e_a}{activation energy}

\item{e_d}{deactivation energy}

\item{k_b}{Boltzmann's constant}

\item{tmp}{temperature}

\item{tmp_r}{reference temperature}

\item{tmp_h}{half-r temperature}
}
\description{
Internal function: function for temperature-dependent population growth
}
