% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_dyn_set.R
\name{hassel}
\alias{hassel}
\title{Internal function: function for community dynamics}
\usage{
hassel(n, r, r0, interaction, k, z)
}
\arguments{
\item{n}{Population size}

\item{r}{Population growth rate}

\item{r0}{Maximum population growth rate at the optimal temperature}

\item{interaction}{Interaction matrix}

\item{k}{Carrying capacity}

\item{z}{Power exponent in the Hassel model. Beverton-Holt (z = 1), under- (z < 1), or over-compensation (z > 1)}
}
\description{
Internal function: function for community dynamics
}
