% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_mcsimp.R
\name{mcsimp}
\alias{mcsimp}
\title{Simulate metacommunity dynamics}
\usage{
mcsimp(
  n_species = 5,
  n_patch = 5,
  n_warmup = 200,
  n_burnin = 200,
  n_timestep = 1000,
  propagule_interval = NULL,
  carrying_capacity = 100,
  interaction_type = "constant",
  alpha = 0,
  min_alpha = NULL,
  max_alpha = NULL,
  r0,
  a,
  e_a,
  e_d,
  tmp_h,
  scale_factor,
  xy_coord = NULL,
  distance_matrix = NULL,
  dispersal_matrix = NULL,
  boundary_condition = "retain",
  outlet = NULL,
  landscape_size = 10,
  mean_env = 0,
  sd_env = 0,
  spatial_env_cor = FALSE,
  phi = 1,
  p_dispersal = 0.1,
  theta = 1,
  dispersal_interval = 1,
  plot = FALSE
)
}
\arguments{
\item{n_species}{Number of species in a metacommunity.}

\item{n_patch}{Number of patches in a metacommunity.}

\item{n_warmup}{Number of time-steps for warm-up. Default \code{200}.}

\item{n_burnin}{Number of time-steps for burn-in. Default \code{200}.}

\item{n_timestep}{Number of time-steps to be saved. Default \code{1000}.}

\item{propagule_interval}{Time interval for propagule introduction during warm-up. If \code{NULL}, a value of \code{ceiling(n_warmup / 10)} will be used.}

\item{carrying_capacity}{Carrying capacity at each patch. Length should be one or equal to \code{n_patch}. Default \code{100}.}

\item{interaction_type}{Species interaction type. \code{"constant"} or \code{"random"}. \code{"constant"} assumes the unique interaction strength of \code{alpha} for all pairs of species. \code{"random"} draws random numbers from a uniform distribution with \code{min_alpha} and \code{max_alpha}.}

\item{alpha}{Species interaction strength. Enabled if \code{interaction_type = "constant"}. Default \code{0}.}

\item{min_alpha}{Minimum value of a uniform distribution that generates species interaction strength. Enabled if \code{interaction_type = "random"}. Default \code{NULL}.}

\item{max_alpha}{Maximum value of a uniform distribution that generates species interaction strength. Enabled if \code{interaction_type = "random"}. Default \code{NULL}.}

\item{r0}{Maximum population growth rate at the optimal temperature}

\item{a}{Normalization constant}

\item{e_a}{Activation energy}

\item{e_d}{Deactivation energy}

\item{tmp_h}{Temperature at which r is half of r_peak.}

\item{scale_factor}{Scale factor.}

\item{xy_coord}{Each row should correspond to an individual patch, with x and y coordinates (columns). Must be give as a data frame. Defualt \code{NULL}.}

\item{distance_matrix}{Distance matrix indicating distance between habitat patches. If provided, the distance matrix will be used to generate dispersal matrix and to calculate distance decay of environmental correlations. Default \code{NULL}.}

\item{dispersal_matrix}{Dispersal matrix to be used to simulate dispersal process. Override distance_matrix. Default \code{NULL}.}

\item{boundary_condition}{Define boundary condition for dispersal. \code{retain} has not loss, \code{loss} induces net loss out of the network.}

\item{outlet}{Outlet patch id.}

\item{landscape_size}{Length of a landscape on a side. Enabled if \code{dispersal_matrix = NULL}.}

\item{mean_env}{Temperature at each patch. Length should be one or equal to \code{n_patch}.}

\item{sd_env}{Standard deviation of temporal temperature variation at each patch.}

\item{spatial_env_cor}{If \code{TRUE}, spatial autocorrelation in temporal environmental fluctuation is considered. Default \code{FALSE}.}

\item{phi}{Parameter describing the distance decay of spatial autocorrelation in temporal environmental fluctuation. Enabled if \code{spatial_env_cor = TRUE}.}

\item{p_dispersal}{Dispersal probability. Length should be one or equal to \code{n_species}.}

\item{theta}{Dispersal parameter describing dispersal capability of species.}

\item{dispersal_interval}{Interval for dispersal event.}

\item{plot}{If \code{TRUE}, five sample patches and species of \code{df_dynamics} are plotted.}

\item{propagule_seed}{Mean number of propagules}
}
\value{
\code{df_dynamics} data frame containing simulated metacommunity dynamics.

\code{df_species} data frame containing species attributes.

\code{df_patch} data frame containing patch attributes.

\code{df_diversity} data frame containing diversity metrics. alpha and gamma diversities were averaged over the simulation time steps (\code{n_timestep}). beta diversity is calculated as gamma / alpha.
}
\description{
Simulate metacommunity dynamics
}
\section{Reference}{
 see \href{https://github.com/aterui/mcbrnet}{github page} for instruction
}

\examples{
\dontrun{
# not run
mcsimp(n_patch = 5, n_species = 5)
}

}
\author{
Akira Terui, \email{hanabi0111@gmail.com}
}
