% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simprotist-package.R
\docType{package}
\name{simprotist-package}
\alias{simprotist}
\alias{simprotist-package}
\title{simprotist: Metacommunity simulation for protist experiments}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
